feat(SubDAG): given a manifest and an id, get the manifest of the DAG with root id
Added a few tests, so I feel decent about how the algorithm works, (especially since it just iterates through the list of links in less then linear time)

However, the manner in which it's coded up is sort of weird to me. My impulse would be to have everything in one function and not create a separate struct to manage everything, but I wanted to try this out since it seems more golang-y. It definitely needs work/review.

Also, realized that SubDAG does not actually solve our SubInfo problem, SubInfo needs to be able to access size before the SubDAG is created, as the indexes in size are the same indexes in the Nodes list.

Perhaps the real way to do this is to have SubInfo be the main function, and have SubDAG be a specific case of SubInfo.