refactor(newSubDAGConverter): expand to include dag.Info
Add Sizes and Paths to newSubDAGConverter, to allow for the creation of SubInfos, without hampering the ability to create a SubDAG

Add InfoAtIndex, InfoAtId, and InfoAtLabel, to create a SubInfo by index, id, or label. The SubInfo will keep the still relevate sizes and labels in the dag.Info.

Adjust `newSubDAGConverter` to take an index, rather then an id.