feat(dsync): add pinning on push completion
dsync needs to know how to pin. This preflights the pinning request so the remote has a
chance to reject early if pinning isn't supported. This'll prevent entire dags being transferred
only to find out pinning isn't possible.

We can add fine-grained controls over remote pinning on top of this foundation, what's important
here is to bring the pinning api into dsync, and optionally use it when a push has completed.